+++ src/commands/apps-sdkconfig.ts
@@ -12,7 +12,7 @@ import {
 } from "../management/apps";
 import { getOrPromptProject } from "../management/projects";
 import { FirebaseError } from "../error";
-import * as requireAuth from "../requireAuth";
+import { requireAuth } from "../requireAuth";
 import * as logger from "../logger";
 import { promptOnce } from "../prompt";